change temperature to float so the calculation will be more precise

+++ Hardware/Aristocrat.Monaco.Hardware/Services/FanService.cs
@@ -32,8 +32,8 @@
 
         public struct TempToSpeed
         {
-            public uint TemperatureLow;
-            public uint TemperatureHigh;
+            public float TemperatureLow;
+            public float TemperatureHigh;
             public uint SpeedLow;
             public uint SpeedHigh;
         };
